Upgrade texture system to 64-bit (16-bit per channel) RGBA textures
- Added GL_RGBA16 internal format support with automatic fallback to GL_RGBA8
- Updated Texture2DArray to accept uint16_t pixel data via GL_UNSIGNED_SHORT
- Added uploadLayer8() fallback path for 8-bit upload when 16-bit unsupported
- Changed AssetManager pixel storage from vector<uint8_t> to vector<uint16_t>
- Load 16-bit PNGs via stbi_load_16() for high color depth
- Updated procedural fallback generator to produce 16-bit per channel data
- Widened BlockTextures texture layer indices from uint8_t to uint16_t
- Widened FloraTypes::textureLayers and GrassSpeciesDefinition::textureLayer
- Updated GreedyMesher tlayer variable to uint16_t to match new types
- Added getBytesPerPixel() accessor to AssetManager
- Added supportsHighBitDepthTextures() and isHighBitDepth() queries
- All 944 assertions in 99 test cases pass

+++ src/content/flora/GrassFactory.hpp
@@ -10,7 +10,7 @@ struct GrassSpeciesDefinition {
   uint16_t speciesId = 0;
   std::string name;
   uint16_t blockId = 0;
-  uint8_t textureLayer = 0;
+  uint16_t textureLayer = 0;
   std::vector<SoilType> acceptableSoil;
   std::vector<std::string> biomeAffinity;  /// Biome names where this grass spawns.
   float spawnChance = 0.1f;  /// 0.0 - 1.0 probability per eligible surface block
